+++ CarCenterImplementation/Models/Kit.cs
@@ -11,7 +11,7 @@ namespace CarCenterImplementation.Models
         public int Id { set; get; }
         [Required]
         public string KitName { set; get; }
-        public int Cost { set; get; }
+        public decimal Cost { set; get; }
         [ForeignKey("KitId")]
         public virtual List<BuiltCarKit> CarKits { set; get; }
         [ForeignKey("KitId")]
